improving validations

+++ src/app/controllers/ContactController.js
@@ -1,4 +1,5 @@
 const ContactsRepository = require('../repositories/ContactsRepository');
+const isValidUUID = require('../utils/isValidUUID');
 
 class ContactController {
   async index(req, res) {
@@ -11,6 +12,10 @@ class ContactController {
 
   async show(req, res) {
     const { id } = req.params;
+
+    if(!isValidUUID(id)) {
+      return res.status(400).json({ error: 'Invalid contact id' });
+    }
 
     const contact = await ContactsRepository.findById(id);
 
@@ -28,25 +33,43 @@ class ContactController {
       return res.status(400).json({ error: 'Name is required' });
     }
 
-    const contactExists = await ContactsRepository.findByEmail(email);
+    if(category_id && !isValidUUID(category_id)) {
+      return res.status(200).json({ error: 'Invalid category' });
+    }
 
-    if(contactExists) {
-      return res.status(400).json({ error: 'This e-mail is already in use' });
+    if (email) {
+      const contactExists = await ContactsRepository.findByEmail(email);
+
+      if(contactExists) {
+        return res.status(400).json({ error: 'This e-mail is already in use' });
+      }
     }
 
     const contact = await ContactsRepository.create({
       name,
-      email,
+      email: email || null,
       phone,
-      category_id,
+      category_id: category_id || null,
     });
 
-    res.json(contact);
+    res.status(201).json(contact);
   }
 
   async update(req, res) {
     const { id } = req.params;
     const { name, email, phone, category_id } = req.body;
+
+    if(!isValidUUID(id)) {
+      return res.status(400).json({ error: 'Invalid contact id' });
+    }
+
+    if(category_id && !isValidUUID(category_id)) {
+      return res.status(200).json({ error: 'Invalid category' });
+    }
+
+    if(!name) {
+      return res.status(400).json({ error: 'Name is required' });
+    }
 
     const contactExists = await ContactsRepository.findById(id);
 
@@ -54,21 +77,19 @@ class ContactController {
       return res.status(404).json({ error: 'Contact not found' });
     }
 
-    if(!name) {
-      return res.status(400).json({ error: 'Name is required' });
-    }
+    if (email) {
+      const contactByEmail = await ContactsRepository.findByEmail(email);
 
-    const contactByEmail = await ContactsRepository.findByEmail(email);
-
-    if(contactByEmail && contactByEmail.id !== id) {
-      return res.status(400).json({ error: 'This e-mail is already in use' });
+      if(contactByEmail && contactByEmail.id !== id) {
+        return res.status(400).json({ error: 'This e-mail is already in use' });
+      }
     }
 
     const contact = await ContactsRepository.update(id, {
       name,
-      email,
+      email: email || null,
       phone,
-      category_id,
+      category_id: category_id || null,
     });
 
     res.json(contact);
@@ -76,6 +97,10 @@ class ContactController {
 
   async delete(req, res) {
     const { id } = req.params;
+
+    if(!isValidUUID(id)) {
+      return res.status(400).json({ error: 'Invalid contact id' });
+    }
 
     await ContactsRepository.deleteById(id);
